+++ examples/spring-data-dgs/src/main/java/com/blazebit/persistence/examples/spring/data/dgs/config/BlazePersistenceConfiguration.java
@@ -20,7 +20,7 @@ import com.blazebit.persistence.Criteria;
 import com.blazebit.persistence.CriteriaBuilderFactory;
 import com.blazebit.persistence.integration.view.spring.EnableEntityViews;
 import com.blazebit.persistence.spi.CriteriaBuilderConfiguration;
-import com.blazebit.persistence.spring.data.impl.repository.BlazePersistenceRepositoryFactoryBean;
+import com.blazebit.persistence.spring.data.repository.config.EnableBlazeRepositories;
 import com.blazebit.persistence.view.EntityViewManager;
 import com.blazebit.persistence.view.spi.EntityViewConfiguration;
 import org.springframework.beans.factory.config.ConfigurableBeanFactory;
@@ -28,7 +28,6 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.context.annotation.Scope;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.PersistenceUnit;
@@ -39,9 +38,8 @@ import javax.persistence.PersistenceUnit;
  */
 @Configuration
 @EnableEntityViews(basePackages = {"com.blazebit.persistence.examples.spring.data.dgs.view"})
-@EnableJpaRepositories(
-        basePackages = "com.blazebit.persistence.examples.spring.data.dgs.repository",
-        repositoryFactoryBeanClass = BlazePersistenceRepositoryFactoryBean.class)
+@EnableBlazeRepositories(
+        basePackages = "com.blazebit.persistence.examples.spring.data.dgs.repository")
 public class BlazePersistenceConfiguration {
 
     @PersistenceUnit